requirements.txt
matplotlib
nltk
numpy
tqdm
pytrec-eval-terrier
pickle